Replace trie.children.remove(idx) by trie.children.splice(idx, 1)
remove() is not a function on Array.
I don't know the semantic the author was going for:
Did he want to remove the item and shorten the array as with splice()?
(that is the semantic supported by Java collections remove() method).
Or did he want to delete trie.children[idx], leaving an undefined element in the array there ?
I went with the splice() semantic.